Json: Refactor WIP
